+++ src/make-filter.js
@@ -1,5 +1,17 @@
-export default (filter) => (
-  `<input
+const filtersSection = document.querySelector(`.main__filter`);
+
+/* Функция отрисовки контента */
+const render = (root, content) => {
+  root.innerHTML = content;
+};
+
+export {filtersSection};
+export default (filters) => {
+  let content = ``;
+
+  /* Шаблон фильтра */
+  const getfilter = (filter) => (
+    `<input
       type="radio"
       id="filter__${filter.label.toLowerCase()}"
       class="filter__input visually-hidden"
@@ -12,4 +24,13 @@ export default (filter) => (
       class="filter__label">${filter.label} 
       <span class="filter__${filter.label.toLowerCase()}-count">${filter.count}</span>
     </label>`
-);
+  );
+
+  /* Заполнение шаблона */
+  filters.forEach((filter) => {
+    content += getfilter(filter);
+  });
+
+  /* Вывод фильтров на станицу */
+  render(filtersSection, content);
+};
