+++ src/make-task.js
@@ -1,4 +1,16 @@
-export default () => (`<article class="card card--pink card--repeat">
+const boardTasks = document.querySelector(`.board__tasks`);
+
+/* Функция отрисовки контента */
+const render = (root, content) => {
+  root.innerHTML = content;
+};
+
+/* Функция отрисовки карточки */
+export default (count) => {
+  let content = ``;
+
+  /* Шаблон карточки */
+  const cardTemplate = () => (`<article class="card card--pink card--repeat">
             <form class="card__form" method="get">
               <div class="card__inner">
                 <div class="card__control">
@@ -12,26 +24,22 @@ export default () => (`<article class="card card--pink card--repeat">
                     favorites
                   </button>
                 </div>
-
                 <div class="card__color-bar">
                   <svg class="card__color-bar-wave" width="100%" height="10">
                     <use xlink:href="#wave"></use>
                   </svg>
                 </div>
-
                 <div class="card__textarea-wrap">
                   <label>
                     <textarea class="card__text" placeholder="Start typing your text here..." name="text">It is example of repeating task. It marks by wave.</textarea>
                   </label>
                 </div>
-
                 <div class="card__settings">
                   <div class="card__details">
                     <div class="card__dates">
                       <button class="card__date-deadline-toggle" type="button">
                         date: <span class="card__date-status">no</span>
                       </button>
-
                       <fieldset class="card__date-deadline" disabled="">
                         <label class="card__input-deadline-wrap">
                           <input class="card__date" type="text" placeholder="23 September" name="date">
@@ -40,11 +48,9 @@ export default () => (`<article class="card card--pink card--repeat">
                           <input class="card__time" type="text" placeholder="11:15 PM" name="time">
                         </label>
                       </fieldset>
-
                       <button class="card__repeat-toggle" type="button">
                         repeat:<span class="card__repeat-status">no</span>
                       </button>
-
                       <fieldset class="card__repeat-days" disabled="">
                         <div class="card__repeat-days-inner">
                           <input class="visually-hidden card__repeat-day-input" type="checkbox" id="repeat-mo-2" name="repeat" value="mo">
@@ -64,7 +70,6 @@ export default () => (`<article class="card card--pink card--repeat">
                         </div>
                       </fieldset>
                     </div>
-
                     <div class="card__hashtag">
                       <div class="card__hashtag-list">
                         <span class="card__hashtag-inner">
@@ -76,7 +81,6 @@ export default () => (`<article class="card card--pink card--repeat">
                             delete
                           </button>
                         </span>
-
                         <span class="card__hashtag-inner">
                           <input type="hidden" name="hashtag" value="repeat" class="card__hashtag-hidden-input">
                           <button type="button" class="card__hashtag-name">
@@ -86,7 +90,6 @@ export default () => (`<article class="card card--pink card--repeat">
                             delete
                           </button>
                         </span>
-
                         <span class="card__hashtag-inner">
                           <input type="hidden" name="hashtag" value="repeat" class="card__hashtag-hidden-input">
                           <button type="button" class="card__hashtag-name">
@@ -97,18 +100,15 @@ export default () => (`<article class="card card--pink card--repeat">
                           </button>
                         </span>
                       </div>
-
                       <label>
                         <input type="text" class="card__hashtag-input" name="hashtag-input" placeholder="Type new hashtag here">
                       </label>
                     </div>
                   </div>
-
                   <label class="card__img-wrap card__img-wrap--empty">
                     <input type="file" class="card__img-input visually-hidden" name="img">
                     <img src="img/add-photo.svg" alt="task picture" class="card__img">
                   </label>
-
                   <div class="card__colors-inner">
                     <h3 class="card__colors-title">Color</h3>
                     <div class="card__colors-wrap">
@@ -125,7 +125,6 @@ export default () => (`<article class="card card--pink card--repeat">
                     </div>
                   </div>
                 </div>
-
                 <div class="card__status-btns">
                   <button class="card__save" type="submit">save</button>
                   <button class="card__delete" type="button">delete</button>
@@ -133,3 +132,14 @@ export default () => (`<article class="card card--pink card--repeat">
               </div>
             </form>
           </article>`);
+
+  let i = 0;
+
+  while (i < count) {
+    content += cardTemplate();
+    i++;
+  }
+
+  /* Вывод фильтров на станицу */
+  render(boardTasks, content);
+};
